refactor: 이미지 변환 로직 분리
