+++ src/utils/imageConverter.js
@@ -1,0 +1,15 @@
+export const imageToBinary = (imgSrc) => {
+  const byteString = atob(imgSrc.split(',')[1]);
+
+  const ab = new ArrayBuffer(byteString.length);
+  const ia = new Uint8Array(ab);
+
+  for (let i = 0; i < byteString.length; i++) {
+    ia[i] = byteString.charCodeAt(i);
+  }
+  const blob = new Blob([ia], {
+    type: 'image/jpeg',
+  });
+
+  return blob;
+};
